Rename interactive-graph-state.ts to initialize-graph-state.ts, after its primary export
